Remove contention on internal.cache.Cache.mu
Caches are used on read-heavy code paths and most of them contain final
 key->value mappings (final: the values never change for a key).
One mapping is not final, query -> descriptor-ids; these change when
 the schema changes. But in that case we are overwriting an existing
 map entry, which is still cheap.

Full locking for read and write operations is not warranted for these
 conditions.
Using a sync.Map eliminates locking at no elevated cost for writes.

A sync.Map consists of a read-only map with in-place updates plus a
 process for adding/deleting entries via a dirty-map that is a shadow
 copy of the ro-map. The dirty-map will eventually get promoted to the
 ro-map.

sync.Maps are not limited in size. Based on the assumption that
 applications have a finite number of:
1. Golang struct types for reading data into
2. Schema for query inputs
3. Schema for query results
 it should be OK to have _unbound_ (well, bound by ^) memory usage here.